pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "examples.allure-behave"
version = "0.0.0"
description = "Getting started with Allure Behave!"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    # Put the project's production dependencies here:

]

[project.optional-dependencies]
dev = [
    # Put the project's dev dependencies here:
    "allure-behave",
    "flake8",
    "flake8-builtins",
]
